+++ src/vulkan/vulkan_function.cpp
@@ -660,7 +660,7 @@ LibraryVulkan::LibraryVulkan(const DeviceVulkan& dev, bool retainBinary)
 
 void LibraryVulkan::loadFromCache(const void* data, size_t length,
                                   const CompilerOptions& options) {
-  auto& cache = Device::binaryCache();
+  auto& cache = _dev.binaryCache();
   if (!cache.isEnabled()) return;
 
   std::vector<std::vector<unsigned char>> binaries;
@@ -685,7 +685,7 @@ void LibraryVulkan::loadFromCache(const void* data, size_t length,
 
 void LibraryVulkan::saveToCache(const void* data, size_t length,
                                 const CompilerOptions& options) const {
-  auto& cache = Device::binaryCache();
+  auto& cache = _dev.binaryCache();
   if (!cache.isEnabled()) return;
 
   std::vector<unsigned char*> binaries = {